test(storefront): fold url.test.ts into storefront.test.ts
Both files imported the same fixture and tested the same two functions, neither
referencing the other, so buildAppleMusicURL changes meant editing two files.
Eight tests had no equivalent and move across: the uppercase and three-letter
storefront rejections, the empty path, the persisted storefront, the locale
fallback, and the three language cases.

The other ten were already covered, including last-page with a language, which
storefront.test.ts tests against a stored path that already carries a query.
